A default gird for the game
In the file I made a default grid. 0 stands for that a field is empty, 1/2/3/4 stands for how many bombs there are contacted and 100 is a bomb. When the grid is printed out, the 100 is printed as a * so that the grid is aligned. This info I will add as a comment in the file tomorrow.